Fix calculation during xcorr linear interpolation
For each value d of tau_diff, lines 72-3 find some integer k such that

    tau_line[k, 0] < d < tau_line[k + 1, 0],

where tau_line is an array storing the values at which the
cross-correlation has been pre-computed. In essence this finds the
closest two values to d at which xcorr has already been calculated. The
code presumably does this to create more accurate linear interpolation
than just using the endpoints.

The problem was that the code assumed that tau_line was 1-indexed, which
makes sense given that the original version was in matlab. As a result,
the code returned a value of k that was 1 too high. This commit fixes
that calculation.

Given that the endpoints used for the interpolation were slightly off,
it's possible that all previous calculations of cross-correlation were
subtly incorrect.